# Add code generation for all wire types for stream encoding