--- a/gen/field.go
+++ b/gen/field.go
@@ -44,6 +44,7 @@ var reservedIdentifiers = map[string]struct{}{
 	"FromWire": {},
 	"String":   {},
 	"Equals":   {},
+	"Encode":   {},
 }
 
 // fieldGroupGenerator is responsible for generating code for FieldGroups.
@@ -92,6 +93,10 @@ func (f fieldGroupGenerator) Generate(g Generator) error {
 	}
 
 	if err := f.FromWire(g); err != nil {
+		return err
+	}
+
+	if err := f.Encode(g); err != nil {
 		return err
 	}
 
@@ -463,6 +468,98 @@ func (f fieldGroupGenerator) FromWire(g Generator) error {
 				<- end>
 			<end>
 			return nil
+		}
+		`, f, TemplateFunc("constantValuePtr", ConstantValuePtr))
+}
+
+func (f fieldGroupGenerator) Encode(g Generator) error {
+	return g.DeclareFromTemplate(
+		`
+		<$stream := import "go.uber.org/thriftrw/protocol/stream">
+
+		<$v := newVar "v">
+		<$sw := newVar "sw">
+		// Encode serializes a <.Name> struct directly into bytes, without going
+		// through an intermediary type.
+		//
+		// An error is returned if a <.Name> struct could not be encoded.
+		func (<$v> *<.Name>) Encode(<$sw> <$stream>.Writer) error {
+			if err := <$sw>.WriteStructBegin(); err != nil {
+				return err
+			}
+			
+			<$structName := .Name>
+			<range .Fields>
+				<- $fname := goName . ->
+				<- $f := printf "%s.%s" $v $fname ->
+				<$t := typeCode .Type>
+				<- if .Required ->
+					<- if and (not (isPrimitiveType .Type)) (not (isListType .Type)) ->
+						if <$f> == nil {
+							return <import "errors">.New("field <$fname> of <$structName> is required")
+						}
+					<- end>
+						if err := <$sw>.WriteFieldBegin(<$stream>.FieldHeader{ID: <.ID>, Type: <$t>,}); err != nil {
+							return err
+						}
+						if err := <encode .Type $f $sw>; err != nil {
+							return err
+						}
+						if err := <$sw>.WriteFieldEnd(); err != nil {
+							return err
+						}
+				<- else ->
+					<- if isNotNil .Default ->
+						<- $fval := printf "%s%s" $v $fname ->
+						<$fval> := <$f>
+						if <$fval> == nil {
+							<$fval> = <constantValuePtr .Default .Type>
+						}
+						{
+							if err := <$sw>.WriteFieldBegin(<$stream>.FieldHeader{ID: <.ID>, Type: <$t>,}); err != nil {
+								return err
+							}
+							if err := <encodePtr .Type $fval $sw>; err != nil {
+								return err
+							}
+					<- else ->
+						if <$f> != nil {
+							if err := <$sw>.WriteFieldBegin(<$stream>.FieldHeader{ID: <.ID>, Type: <$t>,}); err != nil {
+								return err
+							}
+							if err := <encodePtr .Type $f $sw>; err != nil {
+								return err
+							}
+					<- end>
+							if err := <$sw>.WriteFieldEnd(); err != nil {
+								return err
+							}
+						}
+				<- end>
+
+			<end>
+
+			<if and .IsUnion (len .Fields)>
+				<$fmt := import "fmt">
+				<$count := newVar "count">
+				<$count> := 0
+				<range .Fields ->
+					if <$v>.<goName .> != nil {
+						<$count>++
+					}
+				<end>
+				<if .AllowEmptyUnion>
+					if <$count> > 1 {
+						return <$fmt>.Errorf("<.Name> should have at most one field: got %v fields", <$count>)
+					}
+				<else>
+					if <$count> != 1 {
+						return <$fmt>.Errorf("<.Name> should have exactly one field: got %v fields", <$count>)
+					}
+				<end>
+			<end>
+
+			return <$sw>.WriteStructEnd()
 		}
 		`, f, TemplateFunc("constantValuePtr", ConstantValuePtr))
 }