--- a/gen/list.go
+++ b/gen/list.go
@@ -140,6 +140,53 @@ func (l *listGenerator) Reader(g Generator, spec *compile.ListSpec) (string, err
 	return name, wrapGenerateError(spec.ThriftName(), err)
 }
 
+// Encoder generates a function to encode a list given a stream.Writer
+//
+//     func $name(val []listType, sr *stream.Writer) error {
+//             ...
+//     }
+//
+// And returns its name.
+func (l *listGenerator) Encode(g Generator, spec *compile.ListSpec) (string, error) {
+	name := encoderFuncName(g, spec)
+	err := g.EnsureDeclared(
+		`
+		<$stream := import "go.uber.org/thriftrw/protocol/stream">
+
+		<$listType := typeReference .Spec>
+		<$sw := newVar "sw">
+		<$lh := newVar "lh">
+		<$o := newVar "o">
+		<$k := newVar "k">
+		<$v := newVar "v">
+		<$val := newVar "val">
+		func <.Name>(<$val> <$listType>, <$sw> <$stream>.Writer) error {	
+			<$vt := typeCode .Spec.ValueSpec>
+			<$lh> := <$stream>.ListHeader{
+				Type: <$vt>,
+				Length: len(<$val>),
+			}
+			if err := <$sw>.WriteListBegin(<$lh>); err != nil {
+				return err
+			}
+
+			for _, <$v> := range <$val> {
+				if err := <encode .Spec.ValueSpec $v $sw>; err != nil {
+					return err
+				}
+			}
+			return <$sw>.WriteListEnd()
+		}
+		`,
+		struct {
+			Name string
+			Spec *compile.ListSpec
+		}{Name: name, Spec: spec},
+	)
+
+	return name, wrapGenerateError(spec.ThriftName(), err)
+}
+
 // Equals generates a function to compare lists of the given type
 //
 // 	func $name(lhs, rhs $listType) bool {