@dianale31 2021-06-22
Benchmark tests for stream encoding show significant savings for bytes allocated as well as CPU time:

```
BenchmarkRoundTrip/PrimitiveOptionalStruct/Encode-8  	  337783	      2997 ns/op	     704 B/op	       1 allocs/op
BenchmarkRoundTrip/PrimitiveOptionalStruct/Stream_Encode-8         	 1000000	      1059 ns/op	       0 B/op	       0 allocs/op
BenchmarkRoundTrip/PrimitiveOptionalStruct/Decode-8                	  385268	      3520 ns/op	    1400 B/op	      14 allocs/op
BenchmarkRoundTrip/Graph/Encode-8                                  	  328562	      3665 ns/op	    1705 B/op	      11 allocs/op
BenchmarkRoundTrip/Graph/Stream_Encode-8                           	  911430	      1483 ns/op	       0 B/op	       0 allocs/op
BenchmarkRoundTrip/Graph/Decode-8                                  	  158534	      7009 ns/op	    2776 B/op	      32 allocs/op
BenchmarkRoundTrip/ContainersOfContainers/Encode-8                 	   46626	     25660 ns/op	    1305 B/op	      18 allocs/op
BenchmarkRoundTrip/ContainersOfContainers/Stream_Encode-8          	   83979	     14972 ns/op	       0 B/op	       0 allocs/op
BenchmarkRoundTrip/ContainersOfContainers/Decode-8                 	   20068	     61197 ns/op	   12315 B/op	     164 allocs/op
```