@dianale31 2021-06-22
Benchmarks for before and after adding the new stream implementation:
```
name                                        old time/op    new time/op    delta
RoundTrip/PrimitiveOptionalStruct/Encode-8    2.03µs ± 3%    2.17µs ± 2%    +6.94%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Decode-8    2.99µs ± 6%    3.33µs ± 3%   +11.58%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Stream_Encode-8             1.99µs ±11%
RoundTrip/Graph/Encode-8                      3.62µs ± 7%    3.81µs ± 2%      ~     (p=0.095 n=5+5)
RoundTrip/Graph/Decode-8                      6.00µs ± 2%   11.03µs ± 2%   +83.85%  (p=0.008 n=5+5)
RoundTrip/Graph/Stream_Encode-8                              2.41µs ± 2%
RoundTrip/ContainersOfContainers/Encode-8     23.8µs ± 4%    25.7µs ± 8%    +8.11%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-8     53.8µs ± 2%   147.5µs ± 5%  +173.97%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Stream_Encode-8             16.9µs ± 3%

name                                        old alloc/op   new alloc/op   delta
RoundTrip/PrimitiveOptionalStruct/Encode-8      704B ± 0%      704B ± 0%      ~     (all equal)
RoundTrip/PrimitiveOptionalStruct/Decode-8    1.40kB ± 0%    1.45kB ± 0%    +3.43%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Stream_Encode-8            0.00B 
RoundTrip/Graph/Encode-8                      1.70kB ± 0%    1.71kB ± 0%    +0.06%  (p=0.008 n=5+5)
RoundTrip/Graph/Decode-8                      2.78kB ± 0%    3.56kB ± 0%   +28.25%  (p=0.008 n=5+5)
RoundTrip/Graph/Stream_Encode-8                              0.00B
RoundTrip/ContainersOfContainers/Encode-8     1.30kB ± 0%    1.30kB ± 0%    +0.08%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Decode-8     12.3kB ± 0%    28.6kB ± 0%  +132.41%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Stream_Encode-8              0.00B

name                                        old allocs/op  new allocs/op  delta
RoundTrip/PrimitiveOptionalStruct/Encode-8      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
RoundTrip/PrimitiveOptionalStruct/Decode-8      14.0 ± 0%      15.0 ± 0%    +7.14%  (p=0.008 n=5+5)
RoundTrip/PrimitiveOptionalStruct/Stream_Encode-8               0.00
RoundTrip/Graph/Encode-8                        11.0 ± 0%      11.0 ± 0%      ~     (all equal)
RoundTrip/Graph/Decode-8                        32.0 ± 0%      63.0 ± 0%   +96.88%  (p=0.008 n=5+5)
RoundTrip/Graph/Stream_Encode-8                                 0.00
RoundTrip/ContainersOfContainers/Encode-8       18.0 ± 0%      18.0 ± 0%      ~     (all equal)
RoundTrip/ContainersOfContainers/Decode-8        164 ± 0%       837 ± 0%  +410.37%  (p=0.008 n=5+5)
RoundTrip/ContainersOfContainers/Stream_Encode-8                0.00
```